[Refactor] Move component type decision out of the components
