+++ __tests__/components/App.test.js
@@ -1,6 +1,7 @@
 // Copyright 2019 Stanford University see LICENSE for license
 
 import 'jsdom-global/register'
+import 'isomorphic-fetch'
 import React from 'react'
 import { mount, shallow } from 'enzyme'
 import { MemoryRouter } from 'react-router'
