+++ src/components/editor/property/InputLiteral.jsx
@@ -10,7 +10,9 @@ import {
   findNode, getDisplayValidations, getPropertyTemplate, findErrors,
 } from 'selectors/resourceSelectors'
 import InputValue from './InputValue'
-import { booleanPropertyFromTemplate, defaultLanguageId } from 'Utilities'
+import { defaultLanguageId } from 'Utilities'
+import { booleanPropertyFromTemplate } from 'utilities/propertyTemplates'
+
 import _ from 'lodash'
 
 // Redux recommends exporting the unconnected component for unit tests.
