+++ src/components/editor/property/PropertyTemplateOutline.jsx
@@ -5,7 +5,8 @@ import { connect } from 'react-redux'
 import PropTypes from 'prop-types'
 import shortid from 'shortid'
 import OutlineHeader from './OutlineHeader'
-import { booleanPropertyFromTemplate, isResourceWithValueTemplateRef, resourceToName } from 'Utilities'
+import { isResourceWithValueTemplateRef, resourceToName } from 'Utilities'
+import { booleanPropertyFromTemplate } from 'utilities/propertyTemplates'
 import PropertyComponent from './PropertyComponent'
 import ResourceProperty from './ResourceProperty'
 import { findNode, isExpanded, getPropertyTemplate } from 'selectors/resourceSelectors'
